+++ changelog.md
@@ -4,6 +4,7 @@ Upcoming (TBD)
 Internal
 --------
 * Add test coverage for `client_commands.py`.
+* Add test coverage for `cli_runner.py`.
 
 
 1.74.1 (2026/06/18)